[ 노가영 ] 기능 수정 : 회원가입 캠퍼스 구분 및 검색 가능
