+++ src/components/login/social/GoogleSignIn.js
@@ -9,16 +9,12 @@ function GoogleSignIn({ setModalVisible, setLogin }) {
   const history = useHistory();
   const [modalState, setModalState] = useState(true);
 
-  console.log('Where are you', setModalVisible);
-  console.log('work?', modalState);
-
   return (
     <>
       <GoogleLogin
         clientId="13311386829-vlj3ciu02fu1tqriq8dqo0a3nsm4f90u.apps.googleusercontent.com"
         onSuccess={(googleData) => {
           axios
-
             .post(
               `${PUBLIC_IP}/user/sign-in`,
               {
@@ -35,17 +31,20 @@ function GoogleSignIn({ setModalVisible, setLogin }) {
                 setLogin(true);
                 setModalState(false);
                 setModalVisible(false);
-                //console.log("::",modalState)
-                //console.log(":::", setModalVisible(false))
               }
             })
             .catch((error) => {
-              //setModalState(true);
-              if (error.response?.status === 404) {
-                history.push('/register', {
-                  email: googleData.profileObj.email,
-                  provider: 'google',
-                });
+              switch (error.response?.status) {
+                case 404:
+                  message.warning('회원가입이 되지 않은 사용자입니다. 회원가입 페이지로 넘어갑니다.');
+                  history.push('/register', {
+                    email: googleData.profileObj.email,
+                    provider: 'google',
+                  });
+                  break;
+                case 499:
+                  console.log('body가 비어있는 상태입니다.');
+                  break;
               }
             });
         }}
