+++ src/components/login/social/KakaoSignIn.js
@@ -39,7 +39,7 @@ function KakaoSignIn({ setModalVisible, setLogin }) {
               console.log('error?', error.response.request.status);
               switch (error.response?.request.status) {
                 case 404:
-                  message.error('회원가입이 되지 않은 사용자입니다. 회원가입 페이지로 넘어갑니다.');
+                  message.warning('회원가입이 되지 않은 사용자입니다. 회원가입 페이지로 넘어갑니다.');
                   history.push(`/register`, {
                   email: response.kakao_account.email,
                   provider: 'kakao',
